feat : 카카오 리뷰 관련 크롤링 리팩토링 * 유량 제어, Agent 주기적 변경

+++ src/main/java/org/prography/crawler/kakao/dto/map/info/summary/RoadView.java
@@ -2,10 +2,10 @@ package org.prography.crawler.kakao.dto.map.info.summary;
 
 public record RoadView(
     String panoId,
-    int pan,
-    int tilt,
+    double pan,
+    double tilt,
     Point point,
-    int level
+    double level
 ) {
 
 }
